+++ docs/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - The messages opened with `...` close with their `... done`
+- `configuration check PATH` works with the ignore-configuration-file option
 
 
 ## [0.0.41] - 2026-08-06